refactor(decomp): extract automl → new proximadb-automl crate (TD-DECOMP-15)
Creates crates/horizontal/proximadb-automl and moves prediction.rs (31t),
tuning.rs (51t), workload.rs (2t), optimization.rs (1t) + mod.rs tests (4t)
= 89 tests. AutoMLConfig moved to crate lib.rs; service.rs (0t, depends on
control-tier metrics) stays root-side + re-exports AutoMLConfig. optimization.rs
rewrite: super::service::AutoMLConfig → crate::AutoMLConfig.

Verified: automl crate 89 tests pass; root check + server build + automl-clippy
+ fmt + check_workspace_boundaries + check-layering all green.
