+++ Cargo.toml
@@ -65,6 +65,7 @@ members = [
     "crates/binding/proximadb-catalog-introspection",
     "crates/horizontal/proximadb-recall-bench",
     "crates/horizontal/proximadb-resilience",
+    "crates/horizontal/proximadb-automl",
     "crates/horizontal/proximadb-runtime-common",
     "crates/horizontal/proximadb-serialization",
     "crates/horizontal/proximadb-security",
@@ -266,6 +267,7 @@ proximadb-runtime-common = { path = "crates/horizontal/proximadb-runtime-common"
 proximadb-serialization = { path = "crates/horizontal/proximadb-serialization" }
 proximadb-tls = { path = "crates/horizontal/proximadb-tls" }
 proximadb-resilience = { path = "crates/horizontal/proximadb-resilience" }
+proximadb-automl = { path = "crates/horizontal/proximadb-automl" }
 proximadb-storage-common = { path = "crates/storage/proximadb-storage-common" }
 proximadb-storage-ports = { path = "crates/storage/proximadb-storage-ports" }
 proximadb-storage-traits = { path = "crates/storage/proximadb-storage-traits" }
@@ -382,6 +384,7 @@ proximadb-serialization = { path = "crates/horizontal/proximadb-serialization" }
 proximadb-security = { path = "crates/horizontal/proximadb-security" }
 proximadb-tls = { path = "crates/horizontal/proximadb-tls" }
 proximadb-resilience = { workspace = true }
+proximadb-automl = { workspace = true }
 proximadb-storage-common = { path = "crates/storage/proximadb-storage-common" }
 proximadb-storage-ports = { path = "crates/storage/proximadb-storage-ports" }
 proximadb-storage-traits = { path = "crates/storage/proximadb-storage-traits" }
