+++ crates/horizontal/proximadb-automl/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "proximadb-automl"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "AutoML utilities (tuning, prediction, workload analysis) — extracted from root (TD-DECOMP-15)"
+
+[lib]
+name = "proximadb_automl"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+chrono.workspace = true
+rand.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
