+++ docs/10-quality/td/TD-DECOMP-15-extract-automl.adoc
@@ -1,0 +1,41 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-15: Extract automl → proximadb-automl (new horizontal crate)
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing
+| Severity | Low — behavior-neutral re-export
+| Component | new `crates/horizontal/proximadb-automl`; root `src/automl/mod.rs`
+| Relates to | [[TD-DECOMP-1]] (recipe); [[TD-DECOMP-14]] (resilience — same horizontal pattern)
+|===
+
+== Why
+
+`src/automl/` is a cohesive subsystem (automated tuning, prediction, workload analysis,
+**89 tests**) with only one climb-out: `service.rs` (0 tests) → `crate::metrics` (control-tier
+types). The solution: move the 4 clean files + the mod.rs tests to a new `proximadb-automl`
+crate (horizontal tier); keep `service.rs` at root; move `AutoMLConfig` (a pure config struct)
+to the crate so optimization.rs can reference it.
+
+== What moved (behavior-neutral)
+
+* New `crates/horizontal/proximadb-automl/` (Cargo.toml + lib.rs).
+* Moved: `prediction.rs` (31t), `tuning.rs` (51t), `workload.rs` (2t), `optimization.rs` (1t)
+  + mod.rs test block (4t) = **89 tests**.
+* `AutoMLConfig` struct + `Default` impl moved from `service.rs` → crate `lib.rs`; root
+  `service.rs` now re-exports it (`pub use proximadb_automl::AutoMLConfig;`).
+* `optimization.rs`: `super::service::AutoMLConfig` → `crate::AutoMLConfig`.
+* Root `src/automl/mod.rs`: re-exports from crate + keeps `pub mod service;`.
+* Deps: anyhow, chrono, rand, serde, serde_json, tokio, tracing (all workspace).
+
+== Verification
+
+* `cargo check -p proximadb-automl` green; `nextest` — 89 tests pass.
+* `cargo check -p proximadb` + `check_workspace_boundaries` + `check-layering` + server
+build + automl-clippy + fmt + work-commit-check green.
+
+== Follow-ups
+
+* **Running total once merged: ~541 tests out of root** (452 + 89).
